Testing landscape tool. FMOD implementation starting.
Need to actually watch and read some more for the landscaping tool specifically for creating caves and stuff.

Downloaded FMOD.
Created FMOD project for this project.
Intending to learn FMOD to create some interesting modular audio for this project since audio is a key part of horror.
FMOD:
Assets - the actual sounds.
Events - the things needed to play sounds.
Banks - groups of events that get compiled and loaded to be allow the events to be used.
Timeline sheet - like a DAW with time being the main aspect.
Action sheet - plays the sound no matter where time is not the main aspect.
Logic tracks - like looping - within a timeline sheet.
Parameters - variables

Mixer
Routing - contains buses and other callbacks
Main bus is always the last thing the audio goes through.
Snapshots - basically modifiers to the mixer so like if going underwater could have an underwater snapshot which reduces the volume of various buses which contain different events within them to give the idea that the player is underwater.

Still a lot more to learn especially on the unreal side and also just need to start testing with some sounds to get the hand of it.