Insane Writing. Eye adjustments.
Writing that is encrypted using random letters.
This can be decrypted when looking at and within a certain collision range of the writing.
A widget for the writing character and a widget for the writing itself containing a wrap box.
An actor for the insane writing within the world.
An actor for the look at collision activation and deactivation.

Highlighting and dimming the text when looking at it whilst within a certain collision.
Decrypting and encrypting does in random order as well.
Different encryption each time.
Only thing staying the same between encryption and decryption is the number of letters and spaces.

Going to use this to give the player information when they have looped a couple of times and their mind begins to break.
They won't be able to decrypt it until looping.