Updated project to 5.4 to use gravity direction. Gravity inverter.
In order to make use of the gravity direction as a part of the character movement component updated the project to unreal engine 5.4 instead of 5.3.
This meant removing the FMOD plugins and downloading the 5.4 versions and adding those.

New actor called gravity inverter.
Collision that checks when player enters and exits.
When entered activate anti gravity called.
When exited deactivate anti gravity called.

Anti gravity active bool.
Set gravity direction 1 when anti gravity and -1 when not active. Then set orient rotation to movement and then set to false after 1 second delay with retriggerable delay.
Rotation and movement automatic.