Quantum. Gravity Changer.
2 New actors.
Quantum object
Quantum look detection

Taking functionality from schrodingers cat from 7 sins for the quantum object,
Not using was actor recently rendered since it wasnt working properly.

New actor for gravity changer so not changing the whole gravity location but changing the effect of gravity for all physics bodies and pawns and players.
Modulates gravity my amount with setting gravity not enabled whilst in it but simulating gravity using add force.
Will be used in gravity room to show less gravity in the room that will switch gravity in the loop.