Phone setup. Item folder change. Hand sway independence.
Phone actor created.
Simple mesh created in blender (obviously placeholder.
Simple functionality setup for now.
Need to actually have the functionality for showing time and year.
Animation for holding phone.

Updated functionality for hand sway so that it is hand independent which is mostly going to be used for the stationary hand sway stuff but yea works now.