Journalling starting.
Created journal widget.
Figured out why the journal wasnt taking input.
Fixed this with toggling collision for the main menu and the journal.

Need to finish up the change page spread animations.
Just need to figure out the functionality.
Saved a paint image of the intended process to the folder.

Currently using debug key for opening the journal and closing it.

Need to add functionality for hiding journal when clicking pause when in journal.
Need to add check for if main menu open for opening the journal.