Dynamite setup in inventory component. Inventory swapping items. Inventory canvas borders.  Pickup prompt. Equip item in inventory prompt. Need to figure out how to do inventory and journal prompts.
Dynamite actor created.

Dynamite spawn in inventory component and stored.

Inventory swapping items when an item equipped in the left hand.

Pickup prompt fading in and out when entering range and showing which button to click to pickup item.
Prompt in inventory canvas for which button to click to equip item.