Take Damage sound. Landed change. Player taking no damage but calling take damage will now not do anything. Reset player. Changes to game mode functions. Item duplication avoided. Blood effect for damage.
Take damage sound need to add more to it but good enough for now.
When take damage called but 0 damage passed will not do the main functionality of audio and camera shake.
Player landed if potential damage is 0 then just do a small camera shake.

Player dead body.
Player restarting on death using updated functions for game mode.
Need to really properly setup the functions in the game mode for resetting everything at the start of loops.

Adding item to inventory and attempting item pickup check if the item is already in player inventory. If so we don't add it and instead do a camera shake and currently play the inventory unable to open event at item pickup location.
This avoids item duplication and will allow players to go back to areas to pickup items that they may already have without allowing the player to stock up on the same items. (this will only really apply to the rope and dynamite items but yea).

Using blood effect component on player.
Take damage event making use of this component for doing blood drops.
Updated landing graph for the take damage.
Debug key for taking damage currently using unit vector for the normal of the take damage.