Can't jump whilst move input ignored now. Minecart track mesh with spline mesh components. Stale air.
Jump checks for if ignore move input.

Minecart spline mesh setup for using rails along the spline. Looks a little janky if you do really big gaps between the spline points but think its fine.

Stale air actor which damages the player after a short delay and deals damage every second the player is within it.
Using recoloured fog niagara and changing the scale of that relative to the extent of the box collision for the stale air player detection.

Need to do some polish for the stale air with coughing and stuff.