Lock. Temple Lower Exit blocked off now. Wandering areas placed for now.
New actors for the lock. Lock controls whether the gravity inverter is active or not.
Messy lock icons created in aseprite.

Must go through the upper temple now to finish the game to give purpose for the lock.
Wandering areas placed for now need renaming and probably adjusting but need to work on other stuff.